refactor : #4937 PR 피드백 반영(2) - 기능별 클래스 분리
