+++ src/main/java/step2/NumberCalculator.java
@@ -1,0 +1,13 @@
+package step2;
+
+import java.util.stream.IntStream;
+
+public class NumberCalculator {
+
+        private NumberCalculator() {
+        }
+
+        public static int sum(int[] numbers) {
+                return IntStream.of(numbers).sum();
+        }
+}
