+++ src/test/java/step2/StringAddCalculatorTest.java
@@ -1,3 +1,5 @@
+package step2;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
